[expo-location] Made deprecation messages for SDK41 clearer (#12549)
# Why

I was upgrading to SDK41 for my own project, and noticed the deprecation messages for getPermissionsAsync() and requestPermissionsAsync() in expo-location were slightly confusing. The names of the new methods replacing them were given inconsistently - for example without the "Async" or "()" parentheses at the end.

The lack of "Async", for example, made me try to copy and paste the methods without the "Async" to begin with (i.e. "getForegroundPermissions()" rather than "getForegroundPermissionsAsync()": obviously the former method doesn't exist). By fixing this, giving the full name of the new methods in a consistent manner, the deprecation messages should be easier and quicker to understand.

# How

I just updated the deprecation comments to make the method naming consistent, complete, and correct, then rebuilt the package with yarn.

# Test Plan

I checked that all the location tests passed before and after making these chenges. I ran the location tests as per steps 5 and 6 in "Download and Setup" in https://github.com/expo/expo/blob/master/CONTRIBUTING.md