[android][location] Remove usage of deprecated api (#22653)
# Why
As reported in #20663, we seem to resolving a promise twice. I cannot reproduce this but checking Sentry logs on one of the apps I work on where we are using `expo-location`, I can see that it does occur rarely, under specific conditions. It comes from the `LocationRequestCallbacks` in `requestSingleLocation`. Even though we request only a single update, there seems to be some instances where another is received and we try to resolve again. 

Currently, we create a `LocationRequest`by constructing it but this has been deprecated. The docs also mention that methods like `setInterval` and `setNumUpdates` on this object are `hints` and not guarantees. 

# How
Replaced using the `LocationRequest` constructor with its `Builder`. The [documentation](https://developer.android.com/reference/android/location/LocationRequest.Builder#setMaxUpdates(int)) for the method `setMaxUpdates` says that it will only ever receive the specified number of updates and then be removed. 

# Test Plan
NCL ✅

Because I can't reproduce I can't say for sure this fixes the issue, but at least it removes the deprecated API and the documentation seems firmer on the expected behaviour.