[iOS][location] Add guards to avoid parameter crash in task service and background location (#35477)
# Why

Parameter cleanup:

When passing task options to set up the CLLocationManager the options
struct may contain keys with undefined (from JS) values - which will end
up as NSNull values and crash the app. This can be seen in the
API/BackgroundLocation example in BareExpo where some of the option
fields are undefined.

In BareExpo's Background Location demo screen I'm seeing a crash when we
try to store the output from a new location in our user defaults because
it might contain NSNull values.

# How

- Added fix to read options with defaults and both nil/NSNull checks
- Added methods for sanitizing dictionaries before storing them in the
NSUserDefaults.

# Test Plan

- Run API/BackgroundLocation demo in BareExpo (remove `<MapView` to be
able to run this demo in
`apps/native-component-list/src/screens/Location/BackgroundLocationMapScreen.tsx`)
- Click "Start Tracking"
- Observe that BareExpo crashes when trying to read options in
`EXLocationTaskConsumer.m` for parameters that are NSNull and in
`EXTaskService.m` in `_saveConfigWithDictionary` when storing to
NSUserDefaults.
- Apply fix
- This should no longer crash


# Checklist

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>